requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client for API calls
httpx==0.25.2

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment Configuration
python-dotenv==1.0.0

# Database - MongoDB
pymongo[srv]==4.6.0

# AI and Language Processing
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langgraph==0.5.1
openai==1.93.2

# WhatsApp Integration
twilio==9.6.4
pyngrok==7.0.0

# CLI Interface and Output
typer==0.9.0
rich==13.7.0
InquirerPy==0.3.4

# Logging
loguru==0.7.2

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2025.2

# Async Support
aiofiles==23.2.1

# Image Processing and OCR
Pillow==10.1.0
pytesseract==0.3.10

# Receipt Generation Dependencies
# Using PIL (Pillow) only for receipt generation

# Process Management
psutil==5.9.6

# Development Tools (Optional)
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional Dependencies for Production
gunicorn==21.2.0
python-multipart==0.0.6 